Added new ability to configure the binding address behavior for all network daemons and change the default behavior to bind to INADDR_LOOPBACK instead of INADDR_ANY. THis means that all daemons will default to not being generally reachable outside of the local bbs host, and thus, tightening up security a little bit since there is no other general authentication amongst BBS components.
This behavior can be changed in the Config file and can be set
on a daemon-by-daemon basis by setting the configuration parameter
"<daemon>_BIND_ADDR" to a value "*" (INADDR_ANY) or to a specific
IP address ("192.168.1.1").

Also removed some old, unused configuration parsing code in some
daemons. It appears this code has been superceeded by the
centralized option parsing code, which, is apparently a newer
feature of the BBS. (I had no idea).

Also cleaned up tabs in few spots, especially in the tnc daemon
source.