Start a new tncd from the ground up. I intend to bring back in the original KA9Q code and have it retain a bit more of the abstractions it originally held. This will make a few of the following innovations a bit easier to code:
- UI packet mode for APRS and beacon advertisements

A daemon can connect to a control socket and request the ability to
send and/or receive non-proto UI frames. This will make it possible to
have the BBS emit occasional beacon advertisement packets.

The biggest feature by far, however, will be the ability to participate
in the world-wide APRS network, which uses UI packets exclusively.
With the addition of an "aprsd" daemon, the BBS will be able to service
users who have APRS equipment, including a limited interface into the
BBS using APRS's station-to-station short messaging system.

- AX.25 IP gateway

The BBS will be able to send and receive IP packets over AX.25 and
gateway such packets into the host operating system (via a "tun"
interface). Stations on the BBS's RF channels will be able to use the
BBS as an Internet gateway and the AMPR network at large.

- General mailbox cleanup

The current tncd code has some duplicate and/or dead code for mailbox
setup that is hard to understand. The new code will have just one set
of upcalls between the LAPB layer and the mailbox/TCP socket.