Centralized BBS I/O a bit by reusing common functions. "putchar_socket" is now "puts_socket", which is a much better name for what it does. "print_socket" is now "printf_socket", to reflect that it is indeed formatted. CR/LF output translation is now unified in puts_socket, which is the central place to which printable I/O now flows.
Removed duplicated "common.h" which appears to be a complete copy
of bbslib.h and wasn't doing anyone any good.

The macros "PRINT" and "PRINTF" and "GETS", which live in bbslib but
can only possibly be consumed by the bbs itself, will need to be moved
to the BBS or the I/O routines in the BBS will need to move to bbslib.

It should now be possible to talk to DOS-based BBSes on any
port, be it stdio or TCP/IP, simply by using the new -l flag. At
some point we should consider getting rid of the NL->CR translation
in tncd as it can probably be perfectly handled more appropriately
in the bbs itself.