Fix routing header to conform to W0RLI spec, HEADER2 format.
The routing header (or "R:" line) is a forwarding-loop detection element added
to a message when it leaves the BBS, destined for another BBS. For quite some
time, the BSD version of N0ARY has been using N0ARY's 4.1.1 routing header line
format, which wasn't quite up to spec in several ways.

Newly unearthed code for 4.1.2 was discovered and it shows the Bob appears to
have actually started patches to move towards a more standard format, known as
HEADER2 in the W0RLI forwarding specification.

This change completes that move. (Note: There is a Y2K bug in SunOS 4.1.3's
strftime() which affects this format and which will have to be addressed on
the SPARC branch).