Y2K formatting cleanups.
Clean up and expand date formats to avoid Y2K problems and other ambiguities,
but only in spots of the codebase which are user-facing and not prone to
breaking any API contracts.

Also start removing unused "date.c" files which are scattered about the project
and which, if were used, would be troublesome to fix in the Y2K cleanup
process.