Make sure that CLOCAL is used on BSD/modern OSes.
It appears that at least one TNC on N0ARY-1 toggles its CD lines
when receiving packets in KISS mode. This is interesting but has
no bearing on whether the serial port should be considered "closed"
by the kernel. Make sure to set "CLOCAL" mode so that this signal
is ignored (at least when it comes to operating in KISS mode).