Fix issue where telnet client negotiation screws up process.
Telnet clients connecting to this script will always attempt to
negotiate telnet options with it. This script, however, is not
prepared for such "garbage" to be in the TCP stream. This
was normally not a problem as the script was designed to filter
out these characters using an incantation of "sed".

It appears that "sed" technique I was using is now no longer
working with the current version of FreeBSD. As a result, it has
become impossible to log in to the BBS with this gateway script.

This change uses "tr", instead, to do the job of filtering. However,
the fix is pretty ugly as "tr" is now LOCALE sensitive and will
refuse to run if it is given any binary characters that are outside
of the set of legal characters in the current LC_CTYPE.

The fix on top of this fix is to force the locale to be one that
legitimately has no problem with random sequences of high-bit-set
binary characters in input. "bg_BG.CP1251" just so happens to be
one, so that's what I'm using.