Fix build process when not in src directory.
Fix an issue with the build process that crops up when one runs "gmake"
inside a utility directory, rather than at the top root.

Reminder: the build system supports the ability to build a specific tool or
subset of the system by running "gmake" in a sub-directory. This is handy for
testing changes to a specific part of the system, rapidly, rather than waiting
for the entire build process to run. This bug broke that ability, and this
fix restores it.