Improve rc startup.
Re-code the BSD startup script to conform to recommendations from
FreeBSD's rc startup guide by removing manual processing of startup
command arguments and, instead, providing enough context to rc.subr
so that it can be in charge of command invocation.

This fixes a bug where the BBS is simply unable to start up
automatically on modern FreBBSD systems because they have started
providing specialized "verbs" on the rc script command line (such
as "faststart", which was unrecognized by the script in this
particular case).