Fix a few bugs with startup, notably, the inability to stop!
After testing with FreeBSD's "service" command and noticing that under
the new daemon management one is no longer able to stop the BBS with
the "stop" rc command, ensure that the bbs daemon is started up with
the "remain in foreground" command line flag rather than letting it
daemonize itself. Now that the daemon service is in charge, it's important
that the bbs daemon remains in the foreground so that the daemon service
can manage it and tell when it needs a restart.