Change example daemon and bbs directories.
Change the example BBS_DIR to /home/bbs (from /bbs) and likewise
change all the example daemon locations from /bbs to /home/bbs in the example
configuration file so as to set up any aspiring new sysops up for success
should they fail to really scrutinize the examples and, instead, assume
that they have been set up for a modern BSD installation.

Users of the system are more likely to install the BBS in /home/bbs as that
it the most likely place that "useradd" scripts under BSD (and Linux) will
suggest when creating a home directory for a user.