Convert bbsd from select() over to event library.
Convert BBSD from a manually-coded select() event loop over to my
async event-based library, "alib".

There was some very old code in here which was written to manage daemon
spawning, "monitoring" and "pinging" and which was a bit hard to
reverse engineer. It may not have been converted over correctly.