Make customization easier.
Take a step towards possibly fixing issue #20 and others by ensuring that
files that need to be customized are named in a way such that they won't
conflict with canonical souce distributions and, additionally, update the
message compilation tool so that it can take two message files as input.

From now on, the helpmsg tool will compile the BBS help messages from the
canonical source (`src/bbs/helpmsg/helpmsg.txt`) and from the localized source
(`src/include/local-helpmsg.txt`). Any message macros present in the localized
source will override those in the canonical source. In this way, an end
user can safely fix various informational messages in the source code in a
way such that the changes won't conflict or be lost when performing a
source update.