Fix many database update routines to be crash safe.
Fix issue #24 and many others like it by upgrading the existing
"spool_fopen()", "spool_fclose()" framework and adhering to
better error checking.

Many of the daemons in the BBS have fortunately being using
a centralized "spool_xxx()" framework when updating text database
files, but the framework lacked a rollback feature and even with
it, none of its users were checking for corruption issues during
their write/update cycles.

With this change, the framework now has a "spool_abort()" routine
which can be called when a daemon realizes that it cannot
write an update file correctly. Additionally, the framework now
uses the standard "rename()" system call instead of the
non-atomic "unlink/link" sequence it was using before.