Use AsyncLineBuffer in bbsd. Fix annoying bug!
Convert bbsd to use the new AsyncLineBuffer routines for efficient async
reading of lines from clients.

And, fix a long-standing, hard-to-reproduce bug wherein the daemon
occasionally gets stuck in an infinite loop while walking the client list.

The bug was due to code not properly handling error conditions when attemping
to write the _second_ banner line to a newly checked in process. The unhandled
error led to a use-after-free error which eventually led to a corrupted
process list.