Fix a few socket items and a memory leak in bbsd.
Fix a few potential issues with bbsd and how it handles socket I/O, most
of which have likely been hidden and untriggered due to the fact that
most bbsd clients are local and well behaved.

First, with the addition of the async line reader, a bug existed where a
read() that yielded more than one ASCII line of text would ultimately end
up falling behind in input as the further lines wouldn't be parsed
until further input came in. This was hidden by well-behaved, local I/O
which never yielded more than one line per read.

The fix here is to call async_line_read() in a loop until it yields a
"ASYNC_MORE" status.

Second, incoming client sockets were never set to be non-blocking! Client
input that did not cleanly terminate with a newline would block the entire
daemon until a newline arrived. This too was hidden by well-behaved, local
I/O which never yielded anything less than a full line of text.

The fix here is to make sure that the client socket is set to non-blocking
I/O.

Third, a small memory leak was introduced when I took out calls to
"socket_close()" and replaced them with "close()". This is because
I still left in the call to "socket_accept()", which places the incoming
file descriptor in a kind of internal management list inside the tools
library. In addition, "socket_accept()" has likely never been setting the
SO_LINGER option properly on non SunOS machines due to the fact that
setsockopt() on modern OS variants actually takes a "struct linger"
argument, not an integer argument.

The fix here is to update "socket_accept()" to use the right structure
on modern operating systems, and to introduce a new call,
"socket_accept_nonblock_unmanaged()", which performs all the proper flag
checking, setting, and error checking, and which does not attempt to
"manage" the file descriptor.

The fix also required adding a new C flag, "HAVE_LINGER_STRUCT", to the
configuration system. I have assumed that it is true for all OS variants
except for SUNOS.